+++ cli.js
@@ -119,7 +119,7 @@ let cmpVersions = function(a, b) {
 let updateMessage = function(result) {
     utils.warn('');
     utils.alert(utils.colors.red('Please update @turbonemesis/cli to the latest version ' + result.trim()));
-    utils.alert(utils.colors.red('See what\'s changed https://github.com/steveblue/angular2-rollup/blob/master/CHANGELOG.md'));
+    utils.alert(utils.colors.red('See what\'s changed https://github.com/chase2981/angular2-rollup/blob/master/CHANGELOG.md'));
     utils.alert(utils.colors.white('npm i -g @turbonemesis/cli@latest'));
     utils.warn('');
 }